1.1: localisation interface and partially implemented localisation class.
Summary:
This is the beginning for the localisations functions in OpenAuth, should it ever actually be released.
This interface and the localisation class are just the base, we/I still have to go through the entire project and replace hard coded strings with localisation calls
(eg., Localisation.get("string.node")) and fix everything up, as well as write the default en_US and en_UK translations.

This interface is just the original prototype for the ideas I had on this, so any and all changes or improvements are welcomed.

Test Plan:
 - Finish the implementation.
 - Replace hard-coded strings with the Localisation calls.
 - Recompile from the feature branch and run it on a testing server.

Reviewers: OoO, TakSuyu

CC: Pr0Failure

Maniphest Tasks: T50

Differential Revision: http://phabricator.maio.me/D7